Avoid out of range error when using big numbers in financial transactions.
The database schema allows numbers up (+/-) 999_999.99. But as we are also adding the amount to the Ordergroup#account_balance, we use lower barriers to avoid running in errors when updating the account balance.

So, technically the user has to make 10 times the maximum input to raise an account balance error. This should be sufficient, I hope.